requirements.txt
pillow
sqlite3
sv_ttk
matplotlib